regulators: Add inventory_path to Rainier chassis
Add the new "inventory_path" property to the chassis object in the
Rainier regulators JSON configuration file.

This property defines the D-Bus inventory path associated with the
chassis.  The property is needed to create the regulator sensors that
contain telemetry data.

This property is currently optional, but it is being changed to
required.  Modify the configuration file to have the same value it is
currently defaulting to.

Testing:
* Verified new config file loaded without errors
* Verified power on and off of the chassis without errors
* Verified regulator sensors have correct association to chassis
  inventory path.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>
Change-Id: I96d896f53a359ff65f65fabe54677b9b76433ccf