regulators: Changes for VDD ITC level regulators
Ruby Zgabay made several updates to the regulators JSON configuration
file for Rainier.

The primary changes were to modify the voltage output exponent and
several related registers for the VDD VRMs.  This was done to resolve
the following defect:

  SW536683 : Rainier MFG: Processor VRM PN 00E5495 is failing to post
  Pgood on powerup in ITC fabs

Several other smaller changes requested by Abey Mathew are included in
this update.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>
Change-Id: I22007d4561cea75c97295dfe81be0c01e5880041