ibm-ups: Initial version of UPS application (#1)
Create an initial version of the ibm-ups application.  This application
will eventually monitor an Uninterruptible Power Supply (UPS).

The UPS must be connected to a USB port on the BMC using an IBM "System
Port Converter Cable".  The application will eventually read the UPS
status from the cable and publish it on D-Bus using the UPower
org.freedesktop.UPower.Device interface.

This initial commit contains the following:
* Meson changes to build the new application
* Minimal C++ implementation that publishes a D-Bus object with default
  property values

The long-term BMC strategy for UPS support is to use the full UPower
software package.  However, a short-term solution is required to support
IBM systems.  For this reason, this application is being developed
downstream.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>
Change-Id: I19357c6351ca56ae7a5181656915c6a2534dd7b3