ibm-ups: Add --no-poll command line option
Add a '--no-poll' command line option to the ibm-ups application.  If
specified, the application will not poll the UPS device for current
status.

This option is intended for testing other firmware that is affected by
the UPS status.

By default, the ibm-ups application will periodically poll the UPS
device by reading status bits from the USB cable connection.  This
status information is published to a D-Bus interface.

If '--no-poll' is specified, the application will not update the D-Bus
interface.  The D-Bus interface properties can be manually set to the
desired values for testing using the busctl command.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>
Change-Id: I6fd2d6bc3822ba55cfda5194b605158f276a8779