ibm-ups: Poll UPS device for current status
Add support to the ibm-ups application to poll the UPS device for
current status.

The UPS is connected to the BMC using an IBM System Port Converter
cable.  This cable has a serial connector on the UPS side and a USB
connector on the BMC side.

The following bits are read from the cable:
* UPS On: Indicates whether UPS is present/functional.
* Battery Low: Indicates whether the UPS battery level is low.
* Utility Fail: Indicates whether a utility failure has occurred,
                causing the UPS to provide power to the system.

The UPS is polled once per second.  The same bit values must be read
three times in a row before they are considered valid.  This provides
de-glitching to ignore a transient event that caused invalid data.

Once the bit values are considered valid, the corresponding D-Bus
properties are updated.  These properties will be consumed by PLDM and
the state manager application.

Since the UPS is hot-pluggable, the ibm-ups application must handle the
UPS presence changing at any time.  If an attempt to read from the cable
fails three times in a row, the application will assume the UPS has been
removed.

Tested:
* Polling
  * Verified polling occurs once per second
  * Verified polling does not occur if --no-poll command line option
    specified
* UPS presence
  * Tested where UPS is not present when BMC reaches Ready state
  * Tested where UPS is present when BMC reaches Ready state
  * Tested where UPS is added while BMC is running
  * Tested where UPS is removed while BMC is running
  * Verified device path of the form /dev/ttyUSB* exists if UPS is
    present
* Reading status
  * Verified device path is found when UPS is present
  * Verified character special file at device path can be opened, read,
    and closed
  * Verified that once the device is opened, the resulting file handle
    is reused on subsequent polls
  * Verified that same modem control bits must be read 3 consecutive
    times before they are considered valid (to provide de-glitching)
  * Verified that 3 consecutive read failures cause device to be closed
    since UPS may have been removed
  * Verified the UPS status bits were correctly obtained from the modem
    control bits that were read
    * UPS On
    * Battery Low
    * Utility Fail
* D-Bus properties
  * Verified correct initial property values are set when ibm-ups
    application first starts
  * Verified that D-Bus properties are updated when UPS status changes
  * Verified that D-Bus properties are reset to their initial values if
    the UPS is removed
* PLDM
  * Verified that the PLDM state sensor for the UPS is updated when the
    UPS status changes

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>
Change-Id: I08b6ce0e06143cbafeb7f49f58c9b772e5efb6a3