Reset Everest voltage regulators during BMC boot
Add script that runs on Everest systems to reset several voltage
regulators. This is a hardware-workaround required due to some
unexpected power supply behavior. The script was written and tested by
the hardware team.

Also create a service file that runs this script. The service must run
early in the BMC boot process.  It must run before the following
services:
* UPS monitoring
  * One of the affected regulators powers the USB port used to connect
    to a UPS device.
* Software update
  * One of the affected regulators powers the USB port that can be used
    to perform a code update.

The service must run after the following services:
* Power control service
  * Reads a GPIO to determine if the chassis is powered on.  Publishes
    this information on D-Bus.  The script being run needs to check if
    the chassis is powered on.
* MUX work-around
  * The MUX work-around is required in order to access the regulators.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>