Update script that resets Everest regulators
A script needs to be run on Everest systems to reset several voltage
regulators.  This is a hardware-workaround required due to some
unexpected power supply behavior.

In a previous commit, a simple "dummy" version of the script was merged.
This enabled the corresponding service file and build changes to be
tested and merged.

This commit contains the real script that resets voltage regulators.
The script was written and tested by the hardware team.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>
Change-Id: Ifab419064a802a834e59b39f01678811067e8d44