pmbus: Safely call std::filesystem::canonical() (#28)
* Fixup meson.build formatting

Change-Id: I050a9922a3e8dfc9e37d324a90d786277817e5fe
Signed-off-by: Matt Spinler <spinler@us.ibm.com>

* pmbus: Safely call std::filesystem::canonical()

There are a few different places where CALLOUT_DEVICE_PATH elog metadata
is used that passes in the result of a std::filesystem::canonical()
call.

That function will throw if the path passed in does not exist. Since
that code isn't in a try/catch block, the application will crash.

Fix that by using the version of the function that sets a
std::error_code in an output parameter to indicate failure.  If it does
fail, then just use the original path given to it.

As canonical() was used for a reason, there may be side affects such as
proper error analysis not being done, but there is not much that can be
done about that.

Tested:
The app doesn't crash, and can see:

```
Failed to read sysfs file errno=2 FILENAME=/sys/kernel/debug/pmbus/status0
Could not get canonical path from /sys/bus/i2c/devices/3-006a
Failed to read from the device.
```

And in the event log
```
"CALLOUT_DEVICE_PATH" "/sys/bus/i2c/devices/3-006a"
```

Change-Id: I9ea374dc685cc325b40a1f2bb0c8ae0ce71df650
Signed-off-by: Matt Spinler <spinler@us.ibm.com>

* reset_voltage_regulators: turn off formatters

Turn off flake8 with .flake8-ignore, and turn off black with the
'fmt: off' comment inside the script.

Also fixup import ordering for isort since it should be harmless.

Change-Id: I9383ae8335821102bf48e15aac3d15c1738dd43e
Signed-off-by: Matt Spinler <spinler@us.ibm.com>

---------

Signed-off-by: Matt Spinler <spinler@us.ibm.com>